# apps/orders/openapi.yaml
openapi: 3.0.3
info:
  title: Orders Service API
  description: Order management microservice. Integrates with Products service for validation and pricing.
  version: 1.0.0
paths:
  /orders:
    post:
      summary: Create a new order
      description: Validates products via Products service and creates the order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Order created
        '400':
          description: Invalid product or request
    get:
      summary: List all orders
      responses:
        '200':
          description: List of orders
  /orders/by-user/{userId}:
    get:
      summary: List orders for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of orders for the user
  /orders/{id}:
    get:
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
        '404':
          description: Order not found
  /orders/{id}/status:
    patch:
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusDto'
      responses:
        '200':
          description: Order status updated
        '404':
          description: Order not found
components:
  schemas:
    CreateOrderDto:
      type: object
      required:
        - userId
        - items
      properties:
        userId:
          type: string
          example: user_abc123
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItemDto'
        shippingAddress:
          type: string
          example: 123 Main St, City
    CreateOrderItemDto:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          example: prod_001
        quantity:
          type: integer
          minimum: 1
          example: 2
    UpdateOrderStatusDto:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - PENDING
            - CONFIRMED
            - PROCESSING
            - SHIPPED
            - DELIVERED
            - CANCELLED